Enable preview image per post
I was tired of seeing my avatar when I shared a blog post with
people. I changed the way the <head> and <header> tags are
generated so the preview image can be defined in each post.

Also added image previews for the last 2 posts: tarrey town ost
and ci config for jams.